+++ data_orchestrator/data_validator.py
@@ -1,0 +1,16 @@
+# Libraries
+
+import arcpy
+
+# ========================
+# DataValidator
+# ========================
+
+
+class DataValidator():
+    """
+    ...
+    """
+
+    def __init__(self):
+        pass
