+++ docs/setup/test_setup.py
@@ -1,5 +1,5 @@
 from env_setup import environment_setup
-from input_data import input_n50, input_n100
+from data_orchestrator import input_n50, input_n100
 
 # Importing environment
 environment_setup.main()
