+++ generalization/n10/arealdekke/orchestrator/orchestrator.py
@@ -5,8 +5,8 @@ from arealdekke_class import Arealdekke
 from custom_tools.decorators.timing_decorator import timing_decorator
 from env_setup import environment_setup
 
-from input_data import input_area
-from input_data.input_orchestrator import InputDataOrchestrator
+from data_orchestrator.features import input_area
+from data_orchestrator.input_orchestrator import InputDataOrchestrator
 
 
 @timing_decorator
