+++ generalization/n10/facilities/railways_generalization.py
@@ -9,7 +9,7 @@ from env_setup import environment_setup
 from file_manager import WorkFileManager
 from file_manager.n10.file_manager_facilities import Facility_N10
 from generalization.n10.facilities.railways_attributes import main as update_attributes
-from input_data import input_fkb
+from data_orchestrator import input_fkb
 
 
 def compute_line_azimuth(geom: arcpy.Polyline):
