+++ generalization/n10/hoydepunkt/hoydepunkt.py
@@ -12,7 +12,7 @@ from custom_tools.decorators.timing_decorator import timing_decorator
 from env_setup import environment_setup
 from file_manager import WorkFileManager
 from file_manager.n10.file_manager_hoydepunkt import Hoydepunkt_N10
-from input_data import input_n10
+from data_orchestrator import input_n10
 
 arcpy.env.overwriteOutput = True
