+++ generalization/n10/hoydepunkt_innsjo/hoydepunkt_innsjo.py
@@ -7,7 +7,7 @@ from custom_tools.decorators.timing_decorator import timing_decorator
 from env_setup import environment_setup
 from file_manager import WorkFileManager
 from file_manager.n10.file_manager_arealdekke import Arealdekke_N10
-from input_data import input_n10, input_n100
+from data_orchestrator import input_n10, input_n100
 
 arcpy.env.overwriteOutput = True
