+++ generalization/n10/landForms/hoydetall.py
@@ -20,7 +20,7 @@ from custom_tools.general_tools.append_features import Append_Features
 from env_setup import environment_setup
 from file_manager import WorkFileManager
 from file_manager.n10.file_manager_landforms import Landform_N10
-from input_data import input_n10, input_n50, input_n100, input_roads
+from data_orchestrator import input_n10, input_n50, input_n100, input_roads
 
 # ========================
 # Program
