+++ generalization/n100/river/mst_loop.py
@@ -6,7 +6,7 @@ from custom_tools.general_tools import custom_arcpy
 from custom_tools.general_tools.file_utilities import FeatureClassCreator
 from env_setup import environment_setup
 from file_manager.n100.file_manager_rivers import River_N100
-from input_data import input_n50
+from data_orchestrator import input_n50
 
 
 def main():
