+++ generalization/n100/river/unconnected_river_geometry.py
@@ -7,7 +7,7 @@ import config
 from custom_tools.general_tools import custom_arcpy
 from env_setup import environment_setup
 from file_manager.n100.file_manager_rivers import River_N100
-from input_data import input_n50
+from data_orchestrator import input_n50
 
 
 def main():
