+++ generalization/n100/road/data_preparation.py
@@ -10,7 +10,7 @@ from env_setup import environment_setup
 from file_manager.n100.file_manager_roads import Road_N100
 
 # Importing custom input files modules
-from input_data import input_n50
+from data_orchestrator import input_n50
 
 
 @timing_decorator
