+++ generalization/n100/road/data_preparation_2.py
@@ -2,9 +2,10 @@
 import arcpy
 
 # Importing custom input files modules
-from input_data import input_area, input_building, input_railway, input_road
-from input_data.input_datasets import DatasetNamespace
-from input_data.input_orchestrator import InputDataOrchestrator
+from data_orchestrator.features import input_road
+from data_orchestrator.features import input_area, input_building, input_railway
+from data_orchestrator.datasets import DatasetNamespace
+from data_orchestrator.input_orchestrator import InputDataOrchestrator
 
 from composition_configs import core_config, logic_config, type_defs
 from constants.n100_constants import (
