+++ generalization/n100/road/first_generalization.py
@@ -11,7 +11,7 @@ from env_setup import environment_setup
 from file_manager.n100.file_manager_roads import Road_N100
 
 # Importing custom input files modules
-from input_data import input_n100
+from data_orchestrator import input_n100
 
 
 def main():
