+++ generalization/n250/road/dam.py
@@ -15,7 +15,7 @@ from env_setup import environment_setup
 
 # Importing custom modules
 from file_manager.n250.file_manager_roads import Road_N250
-from input_data import input_n100
+from data_orchestrator import input_n100
 
 data_files = {
     # Stores all the relevant file paths to the geodata used in this Python file
