+++ generalization/n250/road/data_preparation_2.py
@@ -37,8 +37,8 @@ from generalization.n250.road.roundabouts import generalize_roundabouts
 from generalization.n250.road.vegsperring import remove_roadblock
 
 # Importing custom input files modules
-from input_data import input_n100, input_roads
-from input_data.input_symbology import SymbologyN250
+from data_orchestrator import input_n100, input_roads
+from data_orchestrator.features.input_symbology import SymbologyN250
 
 MERGE_DIVIDED_ROADS_ALTERATIVE = False
