+++ generalization/n250/road/vegsperring.py
@@ -6,7 +6,7 @@ arcpy.env.overwriteOutput = True
 # Importing custom modules
 from custom_tools.decorators.timing_decorator import timing_decorator
 from file_manager.n250.file_manager_roads import Road_N250
-from input_data import input_n100
+from data_orchestrator import input_n100
 
 data_files = {
     # Stores all the relevant file paths to the geodata used in this Python file
